+++ R/convertToPackage.R
@@ -318,7 +318,7 @@ DESCRIPTIONfileFromModule <- function(module, md, deps, hasNamespaceFile, NAMESP
   d$Imports[hasVersionNumb] <- paste(d$Imports[hasVersionNumb], inequality)
 
   dFile <- filenameFromFunction(packageFolderName, "DESCRIPTION", fileExt = "")
-  origDESCtxt <- if (file.exists(dFile)) read.dcf(dFile) else character()
+  origDESCtxt <- if (file.exists(dFile)) read.dcf(dFile) else character
 
   cat(paste("Package:", d$Package), file = dFile, sep = "\n")
   cat(paste("Type:", d$Type), file = dFile, sep = "\n", append = TRUE)
@@ -328,8 +328,7 @@ DESCRIPTIONfileFromModule <- function(module, md, deps, hasNamespaceFile, NAMESP
   cat(paste("Date:", d$Date), file = dFile, sep = "\n", append = TRUE)
   cat(c("Authors@R:  ", format(d$Authors)), file = dFile, sep = "\n", append = TRUE)
 
-  if (length(d$Imports) || length(origDESCtxt))
-    mergeField(origDESCtxt = origDESCtxt, field = d$Imports, fieldName = "Imports", dFile)
+  mergeField(origDESCtxt = origDESCtxt, field = d$Imports, fieldName = "Imports", dFile)
 
   suggs <- c('knitr', 'rmarkdown', 'testthat', 'withr', 'roxygen2')
   if (length(suggs) || length(origDESCtxt))